+++ CHANGELOG.md
@@ -8,6 +8,7 @@ wlambda code, so you can evaluate code while you are evaluating code.
 * Feature: Thread communication helpers implemented for connecting
 WLambda EvalContext instances accross threads and providing easy
 cross thread messaging and RPC. (TODO: async send).
+* Feature: Custom userdata can be implemented using the VValUserData trait.
 
 0.2.2 (2019-09-01)
 ==================